Analytics_LM/requirements.txt
purecloudplatformclientv2>=17.0.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
requests>=2.31.0
fastapi>=0.100.0
uvicorn>=0.23.0
sqlalchemy>=2.0.0
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.6
plotly>=5.18.0
dash>=2.14.0
dash-bootstrap-components>=1.5.0 